fix: rename i64 -> i32 and f64 -> f32

+++ src/lexer/mod.rs
@@ -200,10 +200,10 @@ impl Lexer<'_> {
         let lexeme = self.in_src[self.start..self.current].to_string();
 
         if is_float {
-            let literal = LiteralValue::NumberFloat(lexeme.parse::<f64>().unwrap());
+            let literal = LiteralValue::NumberFloat(lexeme.parse::<f32>().unwrap());
             self.add_token(TokenType::NumberFloat, lexeme, literal);
         } else {
-            let literal = LiteralValue::NumberInt(lexeme.parse::<i64>().unwrap());
+            let literal = LiteralValue::NumberInt(lexeme.parse::<i32>().unwrap());
             self.add_token(TokenType::NumberInt, lexeme, literal);
         }
     }
